Made attachment honor shadow casting setting.

+++ DiligentEngine.RT/RtStructures.cs
@@ -8,8 +8,8 @@ namespace DiligentEngine.RT
 {
     public class RtStructures
     {
-        public const int OPAQUE_GEOM_MASK = 0x01;
-        public const int TRANSPARENT_GEOM_MASK = 0x02;
+        public const byte OPAQUE_GEOM_MASK = 0x01;
+        public const byte TRANSPARENT_GEOM_MASK = 0x02;
 
         public const int HIT_GROUP_STRIDE = 2;
         public const int PRIMARY_RAY_INDEX = 0;
